+++ src/types/components/CollapseProps.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from 'react';
+
+export type CollapseProps = {
+  isOpen: boolean;
+  children: ReactNode;
+  className?: string;
+};
